OSM gets its extent from bounds OR from params.
When exploring using the OSM datasource to replace postgis in my environment,
I discovered that while postgis takes an extent param, the osm plugin does not.
Since this is a rather useful thing to specify if you're making custom maps,
I figured that I would add this functionality for myself and see if it would
help anyone else out.